Split commands, remove APILink entirely
