--- EGG9000.APILinkSite/.config/dotnet-tools.json
@@ -1,12 +1,0 @@
-{
-  "version": 1,
-  "isRoot": true,
-  "tools": {
-    "dotnet-ef": {
-      "version": "5.0.2",
-      "commands": [
-        "dotnet-ef"
-      ]
-    }
-  }
-}
