--- EGG9000.APILinkSite/Models/ErrorViewModel.cs
@@ -1,7 +1,0 @@
-namespace EGG9000.APILinkSite.Models {
-    public class ErrorViewModel {
-        public string RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
-}
