--- EGG9000.APILinkSite/Program.cs
@@ -1,38 +1,0 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using NLog;
-using NLog.Web;
-using System;
-
-namespace EGG9000.APILinkSite {
-    public class Program {
-        public static void Main(string[] args) {
-            var logger = NLog.LogManager.Setup().LoadConfigurationFromAppSettings().GetCurrentClassLogger();
-            logger.Debug("init main");
-            try {
-                CreateHostBuilder(args).Build().Run();
-            } catch(Exception exception) {
-                // NLog: catch setup errors
-                logger.Error(exception, "Stopped program because of exception");
-                throw;
-            } finally {
-                // Ensure to flush and stop internal timers/threads before application-exit (Avoid segmentation fault on Linux)
-                NLog.LogManager.Shutdown();
-            }
-
-
-        }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureLogging((hostingContext, logging) => {
-                    logging.ClearProviders();
-                    //logging.AddNLog();
-                })
-                .UseNLog()
-                .ConfigureWebHostDefaults(webBuilder => {
-                    webBuilder.UseStartup<Startup>().UseUrls("http://::5014;http://0.0.0.0:5014");
-                });
-    }
-}
