--- EGG9000.APILinkSite/Properties/PublishProfiles/DEV9002ApiLinkSite.pubxml
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-  <PropertyGroup>
-    <WebPublishMethod>MSDeploy</WebPublishMethod>
-    <LaunchSiteAfterPublish>true</LaunchSiteAfterPublish>
-    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
-    <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish />
-    <ExcludeApp_Data>false</ExcludeApp_Data>
-    <ProjectGuid>d45c06f3-515e-4daf-80ad-7817bca3e460</ProjectGuid>
-    <SelfContained>false</SelfContained>
-    <MSDeployServiceURL>https://DEV9002ApiLinkSite.sglade.com</MSDeployServiceURL>
-    <DeployIisAppPath>DEV9002ApiLinkSite</DeployIisAppPath>
-    <RemoteSitePhysicalPath />
-    <SkipExtraFilesOnServer>true</SkipExtraFilesOnServer>
-    <MSDeployPublishMethod>WMSVC</MSDeployPublishMethod>
-    <EnableMSDeployBackup>true</EnableMSDeployBackup>
-    <EnableMsDeployAppOffline>true</EnableMsDeployAppOffline>
-    <UserName>sglade</UserName>
-    <_SavePWD>true</_SavePWD>
-    <_TargetId>IISWebDeploy</_TargetId>
-  </PropertyGroup>
-</Project>
