--- EGG9000.APILinkSite/Properties/PublishProfiles/aws1.sglade.com.pubxml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-This file is used by the publish/package process of your Web project. You can customize the behavior of this process
-by editing this MSBuild file. In order to learn more about this please visit https://go.microsoft.com/fwlink/?LinkID=208121. 
--->
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <WebPublishMethod>MSDeploy</WebPublishMethod>
-    <LaunchSiteAfterPublish>true</LaunchSiteAfterPublish>
-    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
-    <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish />
-    <ExcludeApp_Data>false</ExcludeApp_Data>
-    <ProjectGuid>d45c06f3-515e-4daf-80ad-7817bca3e460</ProjectGuid>
-    <SelfContained>false</SelfContained>
-    <MSDeployServiceURL>egg9000apilinksite.sglade.com</MSDeployServiceURL>
-    <DeployIisAppPath>EGG9000 API Link Site</DeployIisAppPath>
-    <RemoteSitePhysicalPath />
-    <SkipExtraFilesOnServer>false</SkipExtraFilesOnServer>
-    <MSDeployPublishMethod>WMSVC</MSDeployPublishMethod>
-    <EnableMSDeployBackup>true</EnableMSDeployBackup>
-    <UserName>sglade</UserName>
-    <_SavePWD>true</_SavePWD>
-    <AllowUntrustedCertificate>true</AllowUntrustedCertificate>
-    <TargetFramework>net9.0</TargetFramework>
-    <EnableMsDeployAppOffline>true</EnableMsDeployAppOffline>
-    <RuntimeIdentifier>win-x86</RuntimeIdentifier>
-  </PropertyGroup>
-</Project>
