--- EGG9000.APILinkSite/Views/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using EGG9000.APILinkSite
-@using EGG9000.APILinkSite.Models
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
