--- EGG9000.APILinkSite/nlog.config
@@ -1,65 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<!-- XSD manual extracted from package NLog.Schema: https://www.nuget.org/packages/NLog.Schema-->
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd" xsi:schemaLocation="NLog NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      autoReload="true"
-      internalLogFile="c:\logs\egg9000.bot-internal.log"
-      internalLogLevel="Info" >
-
-  <extensions>
-    <add assembly="NLog.Targets.Syslog" />
-    <add assembly="NLog.Targets.GraylogHttp"/>
-  </extensions>
-  <!-- the targets to write to -->
-  <targets async="true">
-    <!-- write logs to file -->
-    <!--<target xsi:type="File" name="logfile" fileName="c:\logs\egg9000.bot.log"
-            layout="${longdate}|${level}|${logger:shortName=true}|${message} |${all-event-properties} ${exception:format=tostring}" />-->
-    <target xsi:type="ColoredConsole" name="logconsole"
-            layout="${date:format=yyyy-MM-dd HH\:mm\:ss}|${level:uppercase=true:truncate=4}|${logger:shortName=true:padding=15:truncate=15}|${message} ${exception:format=tostring}">
-    </target>
-    <target name="syslog" type="Syslog" layout="${level:uppercase=true:truncate=4}|${logger:shortName=true:padding=15:truncate=15}|${message} ${exception:format=tostring}">
-      <messageCreation>
-        <facility>Local7</facility>
-      </messageCreation>
-      <messageSend>
-        <protocol>TCP</protocol>
-        <tcp>
-          <server>logs3.papertrailapp.com</server>
-          <port>22236</port>
-          <tls>
-            <enabled>true</enabled>
-          </tls>
-        </tcp>
-      </messageSend>
-    </target>
-    <!--<target name="graylog"
-              xsi:type="GraylogHttp"
-              facility="EGG9000.Bot"
-            graylogServer="http://192.168.0.73"
-              graylogPort="12201">
-              --><!--graylogServer="http://graylog.dev.sglade.com/"--><!--
-      --><!-- Location information. --><!--
-      <parameter name="source_method" layout="${callsite}" />
-      <parameter name="source_line" layout="${callsite-linenumber}" />
-
-      <parameter name="test_prop" layout="${event-context:item=test_prop}" />
-    </target>-->
-  </targets>
-
-  <!-- rules to map from logger name to target -->
-  <rules>
-    <logger name="*" minlevel="Information" writeTo="syslog">
-      <filters defaultAction='Ignore'>
-        <when condition="contains('${logger}','EGG9000')" action="Log" />
-        <when condition="contains('${level}','ERROR')" action="Log" />
-        <when condition="contains('${level}','CRITICAL')" action="Log" />
-      </filters>
-    </logger>
-    <logger name="*" minlevel="Trace" writeTo="logfile,logconsole">
-      <filters defaultAction='Log'>
-        <when condition="contains('${message}','Rate limit triggered')" action="Ignore" />
-      </filters>
-    </logger>
-  </rules>
-</nlog>
