+++ EGG9000.Bot/Automated/_UpdaterBase.cs
@@ -55,7 +55,6 @@ namespace EGG9000.Bot.Automated {
         public IConfiguration _configuration;
         public IServiceProvider _provider;
         private UpdaterOptions<T> _options;
-        public APILink _apiLink;
 
         protected Bugsnag.IClient _bugSnag;
         protected ILogger<T> _logger;
@@ -87,7 +86,6 @@ namespace EGG9000.Bot.Automated {
             _logger.LogInformation("Initiating");
             _configuration = provider.GetService<IConfiguration>();
             _client = provider.GetService<DiscordHostedService>();
-            _apiLink = provider.GetService<APILink>();
             _dbContextFactory = provider.GetService<IDbContextFactory<ApplicationDbContext>>();
             Instance = this;
             _bugSnag = provider.GetService<Bugsnag.IClient>();
