+++ EGG9000.Common/Helpers/EventHelpers.cs
@@ -64,7 +64,7 @@ namespace EGG9000.Common.Helpers {
 
         private static async Task<Image> GenerateEventImageAsync(Event customEvent) {
             using var client = new HttpClient();
-            client.DefaultRequestHeaders.Add("authenticationKey", Services.APILink.AUTHENTICATION_KEY);
+            client.DefaultRequestHeaders.Add("authenticationKey", DockerSecretsHelper.BotToken);
 
 #if RELEASE
             var baseUrl = "https://egg9000.com";
