+++ EGG9000.Site/Controllers/MyFarmsController.cs
@@ -34,13 +34,12 @@ using Event = EGG9000.Common.Database.Entities.Event;
 namespace EGG9000.Site.Controllers {
     [Authorize]
     public class MyFarmsController(ILogger<MyFarmsController> logger, UserManager<IdentityUser> userManager, DiscordSocketClient discord,
-        RoleManager<IdentityRole> roleManager, APILink apiLink, ApplicationDbContext db, Bugsnag.IClient bugsnag, IMemoryCache cache, DatabaseCache databaseCache) : Controller {
+        RoleManager<IdentityRole> roleManager, ApplicationDbContext db, Bugsnag.IClient bugsnag, IMemoryCache cache, DatabaseCache databaseCache) : Controller {
 
         private readonly ILogger<MyFarmsController> _logger = logger;
         private readonly ApplicationDbContext _db = db;
         private readonly UserManager<IdentityUser> _userManager = userManager;
         private readonly RoleManager<IdentityRole> _roleManager = roleManager;
-        private readonly APILink _apiLink = apiLink;
         private readonly DiscordSocketClient _discord = discord;
         private readonly Bugsnag.IClient _bugsnag = bugsnag;
         private readonly IMemoryCache _cache = cache;
@@ -136,26 +135,18 @@ namespace EGG9000.Site.Controllers {
 
             foreach(var account in user.EggIncAccounts) {
                 var rawBackup = await ContractsAPI.FirstContact(account.Id);
-                //rawBackups.Add(rawBackup.Backup);
                 var customBackup = new CustomBackup(rawBackup.Backup, account?.Backup ?? null);
-                //var json = JsonSerializer.Serialize(customBackup);
-                //var json = Newtonsoft.Json.JsonConvert.SerializeObject(customBackup);
-                //var customBackupAfterJson = Newtonsoft.Json.JsonConvert.DeserializeObject<CustomBackup>(json);
 
-                //var response = await _apiLink.GetBackup(accounts.Id);
                 Console.WriteLine($"Getting backups for {account.Name}");
                 if(customBackup?.Farms is not null) {
                     account.Backup = customBackup;
                     update = true;
                 }
-                //Console.WriteLine(customBackup.SpaceMissions.Count);
 
-                MyContracts scores = _cache.GetOrCreate($"{account.Id}-MyContracts", entry => {
+                var scores = _cache.GetOrCreate($"{account.Id}-MyContracts", entry => {
                     entry.AbsoluteExpirationRelativeToNow = TimeSpan.FromHours(1);
                     return ContractsAPI.Post<MyContracts, BasicRequestInfo>(new BasicRequestInfo(), account.Id).GetAwaiter().GetResult();
                 });
-
-                //var scores = await ContractsAPI.Post<MyContracts, BasicRequestInfo>(new BasicRequestInfo(), account.Id);
 
                 scoring.Add((account.Id, scores));
             }
