+++ EGG9000.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EGG9000.Site", "EGG9000.Sit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EGG9000.Common", "EGG9000.Common\EGG9000.Common.csproj", "{C1984A6C-8660-424F-8A22-0262B9DDAF70}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EGG9000.APILinkSite", "EGG9000.APILinkSite\EGG9000.APILinkSite.csproj", "{D45C06F3-515E-4DAF-80AD-7817BCA3E460}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestProxy", "TestProxy\TestProxy.csproj", "{4EEE22ED-EC4C-4759-A28D-DAD58EE4834A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B01298DD-2A61-469A-BE83-1A5F7C6BC57D}"
@@ -61,14 +59,6 @@ Global
 		{C1984A6C-8660-424F-8A22-0262B9DDAF70}.DEV9002|Any CPU.Build.0 = DEV9002|Any CPU
 		{C1984A6C-8660-424F-8A22-0262B9DDAF70}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C1984A6C-8660-424F-8A22-0262B9DDAF70}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.DEV9001|Any CPU.ActiveCfg = DEV9001|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.DEV9001|Any CPU.Build.0 = DEV9001|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.DEV9002|Any CPU.ActiveCfg = DEV9002|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.DEV9002|Any CPU.Build.0 = DEV9002|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D45C06F3-515E-4DAF-80AD-7817BCA3E460}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4EEE22ED-EC4C-4759-A28D-DAD58EE4834A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4EEE22ED-EC4C-4759-A28D-DAD58EE4834A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4EEE22ED-EC4C-4759-A28D-DAD58EE4834A}.DEV9001|Any CPU.ActiveCfg = DEV9001|Any CPU
