+++ docker-compose.yml
@@ -9,7 +9,6 @@ x-environment: &default-environment
   ConnectionStrings__Token: ...
   ConnectionStrings__ClientSecret: ...
   ConnectionStrings__BugSnagApiKey: ...
-  ConnectionStrings__APILinkUrl: ...
 services:
   # Blue bot instance
   bot_blue:
